requirements.txt
numpy
pandas
albumentations
torch
torchvision
matplotlib
requests
tqdm
gspread
oauth2client
sklearn
pyyaml
